Moved issues to components dir
