+++ src/main/jsx/app.jsx
@@ -8,8 +8,8 @@ import ProjectLayout from './componenets/layout/project-layout';
 import IssueLayout from './componenets/layout/issue-layout';
 import ProjectList from './componenets/project-list';
 import ProjectForm from './componenets/project-form';
-import IssueList from './componenets/issues/issue-list';
-import IssueForm from './componenets/issues/issue-form';
+import IssueList from './componenets/issue-list';
+import IssueForm from './componenets/issue-form';
 import Home from './componenets/home';
 import Profile from './componenets/profile';
 import Settings from './componenets/settings';
